feat: Add Python script to create calendar vents using Exchange Web Services
The above code adds functionality to create a calendar event using Exchange Web Services (EWS) in Python. The script takes input parameters for the subject and date/time of the event, and then creates the event using the specified parameters.

The script uses the exchangelib library to interact with the EWS API and parse the input date/time string using a regular expression. It then creates a CalendarItem object with the specified subject and date/time, and saves it to the user's calendar.e